Make the completer extensible
This commit make it possible for developers to extend the matching
capabilities of the completer. This is important because there are many
legitimate reasons to want to extend the completer, but few which would
make sense to include at the level of this project, e.g. completing Git
branches or Make recipes.

To extend the completer, the user simply uses the "completes" decorator
provided by the Shell class. For example, the following shows how to
extend the completer to complete Git branches::

    @shell.completes
    def git_branches(word):
        branches = sh("git branch --list {}* 2> /dev/null".format(word)).split()
        try:
            branches.remove("*")
        except ValueError:
            pass
        for branch in branches:
            yield branch

Closes #47.